+++ modules/plugin/mongodb/src/main/java/org/geotools/data/mongodb/MongoFeatureStore.java
@@ -63,7 +63,7 @@ public class MongoFeatureStore extends ContentFeatureStore {
     }
 
     @Override
-    protected int getCountInternal(Query query) throws IOException {
+    protected long getCountInternal(Query query) throws IOException {
         return delegate.getCountInternal(query);
     }
