+++ modules/unsupported/vpf/src/main/java/org/geotools/data/vpf/VPFCovFeatureSource.java
@@ -45,7 +45,7 @@ public class VPFCovFeatureSource extends VPFFeatureSource {
         return new VPFFeatureReader(getState(), featureType);
     }
 
-    protected int getCountInternal(Query query) throws IOException {
+    protected long getCountInternal(Query query) throws IOException {
         return -1; // feature by feature scan required to count records
     }
